+++ spec/rubocop/cop/rspec/expect_in_hook_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-RSpec.describe RuboCop::Cop::RSpec::ExpectInHook do
-  subject(:cop) { described_class.new }
+RSpec.describe RuboCop::Cop::RSpec::ExpectInHook, :config do
+  subject(:cop) { described_class.new(config) }
 
   it 'adds an offense for `expect` in `before` hook' do
     expect_offense(<<-RUBY)
@@ -75,5 +75,37 @@ RSpec.describe RuboCop::Cop::RSpec::ExpectInHook do
         expect_any_instance_of(something).to receive(:foo)
       end
     RUBY
+  end
+
+  context 'with config IgnoreSharedGroups set to false' do
+    let(:cop_config) { { 'IgnoreSharedGroups' => false } }
+
+    it 'does not accepts `expect` in `shared_examples`' do
+      expect_offense(<<-RUBY)
+        shared_examples 'for shared setup' do
+          before do
+            expect(object).to receive(:message)
+            ^^^^^^ Do not use `expect` in `before` hook
+            expect_any_instance_of(something).to receive(:foo)
+            ^^^^^^^^^^^^^^^^^^^^^^ Do not use `expect_any_instance_of` in `before` hook
+          end
+        end
+      RUBY
+    end
+  end
+
+  context 'with config IgnoreSharedGroups set to true' do
+    let(:cop_config) { { 'IgnoreSharedGroups' => true} }
+
+    it 'accepts `expect` in `shared_examples`' do
+      expect_no_offenses(<<-RUBY)
+      shared_examples 'for shared setup' do
+        before do
+          expect(object).to receive(:message)
+          expect_any_instance_of(something).to receive(:foo)
+        end
+      end
+      RUBY
+    end
   end
 end